back/requirements.txt
Django==1.9.2
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-filter==0.12.0
django-rest-auth==0.7.0
djangorestframework==3.3.2
gunicorn==19.4.5
Markdown==2.6.5
mod-wsgi==4.4.22
mysqlclient==1.3.7
six==1.10.0
virtualenv==14.0.5
wheel==0.26.0
